% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_model_functions.R
\name{dbetamp}
\alias{dbetamp}
\alias{pbetamp}
\alias{qbetamp}
\alias{rbetamp}
\title{The Beta Distribution Using Mean and Precision}
\usage{
dbetamp(x, mean, prec, log = FALSE)

pbetamp(q, mean, prec, lower.tail = TRUE, log.p = FALSE)

qbetamp(p, mean, prec, lower.tail = TRUE, log.p = FALSE)

rbetamp(n, mean, prec)
}
\arguments{
\item{x, q}{Vector of quantiles
(\code{0 >=} \code{x},\code{q} \code{<= 1}).}

\item{mean}{Vector of means for the beta distribution, where
\code{0 > mean > 1}.}

\item{prec}{Vector of dispersion parameters for the beta
distribution, where \code{prec > 0}.}

\item{log, log.p}{Logical; if \code{TRUE}, probabilities p
are given as log(p).}

\item{lower.tail}{Logical; if \code{TRUE} (default), probabilities
are P(X less than or equal to x) otherwise, P(X greater than x).}

\item{p}{Vector of probabilities}

\item{n}{Number of observations. If \code{length(n) > 1}, the
length is taken to be the number required.}
}
\value{
\code{dbetamp} gives the density, \code{pbetamp} gives the
distribution function, \code{qbetamp} gives the quantile function,
and \code{rbetamp} generates random deviates. The length of the
result is determined by \code{n} for \code{rbetamp}, and is the
maximum of the lengths of the numerical arguments for the other
functions.
}
\description{
Density, distribution function, quantile function, and random
generation for the beta distribution with mean equal to
\code{mean} and precision equal to \code{prec} (lower values
correspond to higher variance).
}
\details{
The mean and precision for the beta distribution can be converted
back to the standard positive shape parameters via
\code{shape1 = mean x prec} and \code{shape2 = (1-mean) x prec}.

See \link[stats:Beta]{stats::dbeta} for more details.
}
\examples{
x <- c(.25, .5, .75)
dbetamp( x, .5, 4)
# Same as 'dbeta' function
dbetamp( x, .5, 4) == dbeta( x, 2, 2)
pbetamp( x, .5, 4)
qbetamp( pbetamp( x, .5, 4), .5, 4)
rbeta( 3, .5, 4)

}
