% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd-package.R
\docType{package}
\name{ddabcd-package}
\alias{ddabcd-package}
\title{Delay Discounting Models for the ABCD study}
\description{
The \pkg{ddabcd} package provides assorted
functions to help with processing and
modeling delay discounting data collected
by the \href{https://abcdstudy.org/}{ABCD study}. See
\url{https://rettopnivek.github.io/ddabcd/}
for more details.
}
