% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_data.R
\name{ddabcd_data_censor}
\alias{ddabcd_data_censor}
\title{Function to Left and Right-Censor Data}
\usage{
ddabcd_data_censor(
  obj_data,
  chr_outcome = "ddabcd_indifference_score_1",
  lgc_indicator = FALSE,
  num_limits = c(0.78125, 99.21875),
  num_scaling = 100
)
}
\arguments{
\item{obj_data}{Either a vector of values
or a data frame.}

\item{chr_outcome}{An optional character string,
the column in \code{obj_data} to left and
right-censor if \code{obj_data} is a data frame.}

\item{lgc_indicator}{Logical; if \code{TRUE}
returns an indicator for whether values
were censored (-1 for left-censored, 1 for
right-censored, and 0 otherwise) rather than
the censored data itself.}

\item{num_limits}{A numeric vector with the lower
and upper limits below and above which to
censor values, respectively. By default is
set to the lowest and highest possible
indifference scores found in the
\href{https://abcdstudy.org/}{ABCD study} data set.}

\item{num_scaling}{A numeric value, a scaling
constant by which to adjust the numeric values
(useful, for instance, if data is in the
range of 0 - 1 but the limits are in the range
of 0 - 100).}
}
\value{
Either a numeric vector of censored value,
or if \code{lgc_indicator} is \code{TRUE}, an
integer vector with values of -1 for left-censored
data, 1 for right-censored data, and 0 otherwise.
If \code{obj_data} is \code{NULL}, returns the
default limits used in the ABCD study.
}
\description{
Function to left and right-censor numeric
values in a column in a data frame.
}
\examples{
# Example data set
dtf_data <-
  data.frame( ddabcd_indifference_score_1 = c( 0, 1, .5 ) )

# Indicator for censored values
ddabcd_data_censor( dtf_data, lgc_indicator = TRUE )
# Censored values
ddabcd_data_censor( dtf_data )

# Vector of values
ddabcd_data_censor(
  c(-5, 0, 5) \%transform\% 'logistic', chr_outcome = ''
)

}
