% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_data.R
\name{ddabcd_data_default_col}
\alias{ddabcd_data_default_col}
\title{Default Column Names for ABCD Study Delay Discounting Data}
\usage{
ddabcd_data_default_col(lgc_long = TRUE)
}
\arguments{
\item{lgc_long}{Logical; if \code{TRUE} returns
column names for the long-form data set, otherwise
returns column names for the wide-form data set.}
}
\value{
A character vector.
}
\description{
Convenience function that returns the default column names
used by the ABCD study and the \code{ddabcd} package
for delay discounting variables.
}
