% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_data.R
\name{ddabcd_data_delay_durations}
\alias{ddabcd_data_delay_durations}
\title{Delay Durations for the ABCD Study Delay Discounting Task}
\usage{
ddabcd_data_delay_durations(chr_units = "log months")
}
\arguments{
\item{chr_units}{A character string, either \code{'months'}
or \code{'log months'}, the units for the delay durations.}
}
\value{
A numeric vector.
}
\description{
Convenience function that returns the 7 delay durations
in order used by the ABCD study.
}
