% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_data.R
\docType{data}
\name{ddabcd_data_example_long}
\alias{ddabcd_data_example_long}
\title{Example Long-Form Data}
\format{
\subsection{\code{ddabcd_data_example_long}}{

A data frame with 672 rows and 21 columns:
\describe{
\item{src_subject_id}{
Participant ID
}
\item{eventname}{
Study time point
}
\item{age}{
Participant age in months
}
\item{sex}{
Biological sex at birth for participant
}
\item{ddis_scr_val_immedcho}{
Count of times participant picked rational immediate choice for 3
validity check questions
}
\item{ddis_scr_val_immedcho}{
Count of times participant picked rational immediate choice for 3
validity check questions
}
\item{ddabcd_simulation_lnk}{
The generating parameter value for the participant
(log discounting rate)
}
\item{ddabcd_simulation_s}{
The generating parameter value for the participant
(sensitivity)
}
\item{ddabcd_simulation_lnd}{
The generating parameter value for the participant
(log standard deviation for logit-normal distribution)
}
\item{ddacbd_quality_no_missing_data}{
Indicator of non-NA indifference scores
}
\item{ddabcd_quality_picked_rational_immediate_choices}{
Indicator of when participant picked all 3 rational choices
for validity check questions
}
\item{ddabcd_quality_variation_in_responses}{
Indicator for when participant did not have the same
indifference score across all delay durations
}
\item{ddabcd_quality_met_Johson_and_Bickel_2008_criteria }{
Indicator for when participant's data met criteria of
Johnson and Bickel (2008)
}
\item{ddabcd_quality_significant_Mann_Kendall_test}{
Indicator for when participant's data had a significant
one-sided Mann-Kendall test of monotonicity
}
\item{ddabcd_delay_columns}{
The original column names for the indifference scores across
different delay durations
}
\item{ddabcd_indifference_score_100}{
Indifference scores ranging from 0 - 100
}
\item{ddabcd_delay_months}{
Delay durations in months
}
\item{ddabcd_delay_log_months}{
The log of the months for delay durations
}
\item{ddabcd_delay_negative_log_months}{
The negative of the log of the months for delay durations
}
\item{ddabcd_indifference_score_1}{
Indifference scores ranging from 0 - 1
}
\item{ddabcd_indifference_score_log_odds}{
Indifference scores ranging from 0 - 1 converted to log-odds
}
\item{ddabcd_indifference_score_censored}{
Indicator for when an indifference score (ranging from 0 - 100)
was left-censored (value fixed to minimum of 0.78125) with a score
of -1, right-censored (value fixed to maximum of 99.21875) with a
score of 1, or not censored with a score of 0
}
}
}
}
\usage{
ddabcd_data_example_long
}
\description{
An example data set with simulated indifference
scores for 48 participants over two time points
with separate rows for each delay duration, time
point, and participant. Data are the result of the
application of the \link{ddabcd_data_prep}
function to the \link{ddabcd_data_example_wide}
data set. Values were simulated from Rachlin's (2006)
\href{https://doi.org/10.1901/jeab.2006.85-05}{model}
with data following a logit-normal distribution.
The generating log discounting rate and sensitivity
parameters were specified to be weakly correlated with
biological sex.
}
\keyword{datasets}
