% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_data.R
\docType{data}
\name{ddabcd_data_example_wide}
\alias{ddabcd_data_example_wide}
\title{Example Wide-Form Data in Format of ABCD Study}
\format{
\subsection{\code{ddabcd_data_example_wide}}{

A data frame with 96 rows and 15 columns:
\describe{
\item{src_subject_id}{
Participant ID
}
\item{eventname}{
Study time point
}
\item{age}{
Participant age in months
}
\item{sex}{
Biological sex at birth for participant
}
\item{ddis_scr_val_indif_pnt_1da}{
Indifference score for delay of 1 day
}
\item{ddis_scr_val_indif_pnt_1week}{
Indifference score for delay of 1 day
}
\item{ddis_scr_val_indif_pnt_1mth}{
Indifference score for delay of 1 month
}
\item{ddis_scr_val_indif_pnt_3mth}{
Indifference score for delay of 3 months
}
\item{ddis_scr_val_indif_pnt_1yr}{
Indifference score for delay of 1 year
}
\item{ddis_scr_val_indif_pnt_5yr}{
Indifference score for delay of 5 years
}
\item{ddis_scr_val_immedcho}{
Count of times participant picked rational immediate choice for 3
validity check questions
}
\item{ddabcd_simulation_lnk}{
The generating parameter value for the participant
(log discounting rate)
}
\item{ddabcd_simulation_s}{
The generating parameter value for the participant
(sensitivity)
}
\item{ddabcd_simulation_lnd}{
The generating parameter value for the participant
(log standard deviation for logit-normal distribution)
}
}
}
}
\usage{
ddabcd_data_example_wide
}
\description{
An example data set with simulated indifference
scores for 48 participants over two time points
formatted in the same style as data from the the
\href{https://abcdstudy.org/}{ABCD study}. Values
were simulated from Rachlin's (2006)
\href{https://doi.org/10.1901/jeab.2006.85-05}{model}
with data following a logit-normal distribution.
The generating log discounting rate and sensitivity
parameters were specified to be weakly correlated with
biological sex.
}
\keyword{datasets}
