% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_data.R
\name{ddabcd_data_merge_by_id_and_time_point}
\alias{ddabcd_data_merge_by_id_and_time_point}
\title{Add Columns to a Data Frame by Participant ID and Time Point}
\usage{
ddabcd_data_merge_by_id_and_time_point(
  dtf_main,
  dtf_to_add,
  chr_participant = "src_subject_id",
  chr_time_point = "eventname",
  num_default = NA
)
}
\arguments{
\item{dtf_main}{A data frame to which columns should be added.}

\item{dtf_to_add}{A data frame from which columns should be taken.}

\item{chr_participant}{A character string, the column for
participant IDs to match over. This column must exist in both
\code{dtf_main} and \code{dtf_to_add}.}

\item{chr_time_point}{An optional character string, the column
for time points to match over. If specified, this column must
exist in both \code{dtf_main} and \code{dtf_to_add}.}

\item{num_default}{The default value to use for missing data.}
}
\value{
A data frame, \code{dtf_main} with additional columns.
}
\description{
Function to add columns from one data frame to another,
matching over participants and time points shared across
data sets. Useful, for example, to combine the different
data sets provided by the \href{https://abcdstudy.org/}{ABCD study}.
}
\examples{
# Example data set
data(ddabcd_data_example_wide)

# Main data set just has ID and time point
dtf_main <- ddabcd_data_example_wide[, 1:2]
# Want to add columns for demographics (age and sex)
dtf_to_add <- ddabcd_data_example_wide[, 1:4]
# Merge only by ID (duplicate values over time points)
dtf_main <- ddabcd_data_merge_by_id_and_time_point(
  dtf_main, dtf_to_add, chr_time_point = ''
)
head( dtf_main, n = 4 )

# Want to add indifference score at delay of 6 hours
# by participant and time point
dtf_to_add <- ddabcd_data_example_wide[, c( 1:2, 5)]
dtf_main <- ddabcd_data_merge_by_id_and_time_point(
  dtf_main, dtf_to_add, chr_time_point = 'eventname'
)
head( dtf_main, n = 4 )

}
