% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_data.R
\name{ddabcd_data_prep}
\alias{ddabcd_data_prep}
\title{Prepare ABCD Data for Delay Discounting Modeling}
\usage{
ddabcd_data_prep(
  dtf_data,
  chr_columns = "",
  chr_indifference = "",
  chr_delay = "",
  chr_quality = "",
  num_delay = NULL,
  chr_immediate_choice_score = "",
  num_alpha_for_MK_test = 0.05
)
}
\arguments{
\item{dtf_data}{A data frame with one row per
each time point and ABCD participant.}

\item{chr_columns}{A character vector with the 7
columns with the indifference score (ranging from
0 - 100) at each delay duration (in order from the
shortest delay to the longest delay).}

\item{chr_indifference}{A character vector with the
4 new column names in the long-form data set
for 1) the indifference score (0 - 100), 2) the
indifference score proportion (0 - 1), 3) the
log-odds of the indifference score proportion,
and 4) an indicator for whether the log-odds were
left or right censored (-1 = left, 1 = right, 0 = none).}

\item{chr_delay}{A character vector with the 4 new
column names in the long-form data set for 1)
the original column names at each delay duration,
2) the delay duration in months, 3) the log of the
delay duration in months, and 4) the negative of
the log of the delay duration in months.}

\item{chr_quality}{A character vector with the 5 new
column names in the long-form data set for
\enumerate{
\item an indicator of non-missing data by participant,
\item an indicator for participants who picked all 3
rational immediate choice responses for the
delay discounting validity check, 3) an
indicator for when participants did not make
the same response across all delay durations,
\item an indicator for when participant data met
the Johnson & Bickel (2008) criteria for
valid data, and 5) an indicator for when
participant data exhibit a statistically
significant monotonic downward trend based
on a one-sided Mann-Kendall test.
}}

\item{num_delay}{An optional numeric vector with
the 7 values to use for the delay duration
(in order from shortest to longest). Otherwise
defaults to duration in months for delays of
6 hours, 1 day, 1 week, 1 month, 3 months,
1 year, and 5 years.}

\item{chr_immediate_choice_score}{A character string
the column name in \code{dtf_data} tracking the
frequency in which participants chose up to 3
rational immediate choice responses during a
validity check.}

\item{num_alpha_for_MK_test}{A proportion, the
false-positive rate for the one-sided Mann-Kendall
test.}
}
\value{
A data frame.
}
\description{
Function to convert wide-form ABCD data with
indifference scores from the delay discounting task
into a long-form data set ready for modeling.
}
\examples{
# Load in example data set
data("ddabcd_data_example_wide")

# Process data set
dtf_delay <- ddabcd_data_prep(ddabcd_data_example_wide)

}
\references{
Johnson, M. W., & Bickel, W. K. (2008). An algorithm
for identifying nonsystematic delay-discounting data.
Experimental and Clinical Psychopharmacology, 16 (3),
264 - 274.
https://doi.org/10.1037/1064-1297.16.3.264
}
