% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_model_functions.R
\name{ddabcd_estimation_priors}
\alias{ddabcd_estimation_priors}
\title{Compute Priors Over Delay Discounting Model Parameters}
\usage{
ddabcd_estimation_priors(num_param, lst_priors, chr_distribution)
}
\arguments{
\item{num_param}{A numeric named vector of parameter values,
where \code{num_param['a0']} is the intercept only (logit-scale),
\code{num_param['lnk']} is for log discounting,
\code{num_param['s']} is for sensitivity, and
\code{num_param['lnd']} is log dispersion.}

\item{lst_priors}{A named list of vectors specifying the
mean and standard deviation, respectively, for the
prior on each parameter.}

\item{chr_distribution}{A character string specifying
the distribution for the data. Options are
\code{'Normal'}, \code{'Beta'}, or \code{'Logit-normal'}.}
}
\value{
A numeric value, a sum of log-likelihoods.
}
\description{
Function to compute priors over delay
discounting model parameters.
}
\details{
The following parameters are given normal priors with
user-specified means and standard deviations:
\itemize{
\item{"a0"}{intercept only on logit-scale}
\item{"lnk"}{log discounting}
\item{"s"}{sensitivity}
\item{"lnd"}{log dispersion}
}

The interpretation of the log dispersion parameter
depends on the distribution for the data. For
normal and logit-normal distributions lower values
indicate lower variance, whereas for the beta
distribution, higher values indicate lower variance.
}
\examples{
# Example normal prior with mean of 0 and
# sd of 1 on intercept on logit-scale
ddabcd_estimation_priors(
  c( a0 = 0 ), lst_priors = list( a0 = c( 0, 1 ) )
)

}
