% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_model_functions.R
\name{ddabcd_estimation_succeeded}
\alias{ddabcd_estimation_succeeded}
\title{Check if Model Estimation Succeeded}
\usage{
ddabcd_estimation_succeeded(obj_fit)
}
\arguments{
\item{obj_fit}{An R object, model fit output.}
}
\value{
A logical value, \code{TRUE} if
\code{obj_fit} is not a character string
(indicating successful fit) or \code{FALSE}
otherwise.
}
\description{
Convenience function that checks the
class of a model fit object. Fail-states
for model fit result in a simple
character string reporting failure rather
than the complex class with results,
allowing confirmation of success if an
object is not a character string.
}
