% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_model_functions.R
\name{ddabcd_estimation_sum_of_log_tobit_likelihoods}
\alias{ddabcd_estimation_sum_of_log_tobit_likelihoods}
\title{Compute Sum of Log Tobit Likelihoods for Delay Discounting Models}
\usage{
ddabcd_estimation_sum_of_log_tobit_likelihoods(
  num_param,
  dtf_data,
  chr_measures = c("ddabcd_indifference_score_log_odds",
    "ddabcd_delay_negative_log_months", "ddabcd_indifference_score_censored"),
  num_limits = NULL,
  num_fnscale = -1,
  lst_priors = NULL,
  fun_priors = NULL
)
}
\arguments{
\item{num_param}{A numeric named vector of parameter values,
where \code{num_param['a0']} is the intercept only (logit-scale),
\code{num_param['nlnk']} is for minus log discounting,
\code{num_param['s']} is for sensitivity, and
\code{num_param['lnd']} is log dispersion.}

\item{dtf_data}{A data frame with the log-odds of the indifference
scores, the minus log of the delay duration, and an indicator
for whether the data are left-censored (-1), right-censored (1),
or uncensored (0).}

\item{chr_measures}{A character vector giving the column names
in \code{dtf_data} for 1) the log-odds of the indifference
scores, 2) the minus log delays, and 3) the censoring indicator.}

\item{num_fnscale}{An numeric value of either 1 or -1
adjusting the sum of the log-likelihoods (by default
is -1 to ensure compatibility with functions like
\link[bbmle:mle2]{bbmle::mle2}).}

\item{lst_priors}{An optional named list with the
hyperparameters for the priors on each parameter.}

\item{fun_prior}{An optional function that takes the
vector of parameters and a list of hyperparameters
to compute a prior on the parameters. Allows for
penalized maximum likelihood estimation.}
}
\value{
A numeric value, the sum of the log-tobit-likelihoods
adjusted by \code{num_fnscale}.
}
\description{
Function to compute the sum of the log-likelihoods for
one of three delay discounting models: 1) an intercept-only
model, 2) Mazur's (1987) model, or 3) Rachlin's (2006) model.
Can be passed to optimization routines like
\link[stats:optim]{stats::optim} or \link[bbmle:mle2]{bbmle::mle2} to conduct maximum likelihood
estimation. The outcome is required to be on the logit scale
and is assumed to be both left and right-censored, and the
tobit likelihood function (CITATION) is used as a result.
}
\examples{
# Examples

}
