% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_model_functions.R
\name{ddabcd_fit_models_across_participants}
\alias{ddabcd_fit_models_across_participants}
\title{Fit Models Across Participants}
\usage{
ddabcd_fit_models_across_participants(
  dtf_data,
  fun_to_fit_models,
  chr_participant = "src_subject_id",
  lgc_progress = TRUE,
  ...
)
}
\arguments{
\item{dtf_data}{A data frame with separate rows for each
delay duration and participant. Must have columns for
the indifference scores (ranging from 0 to 1) and for
log of the delay durations.}

\item{fun_to_fit_models}{A function to fit the delay
discounting models to the data (see
\link{ddabcd_fit_models_using_nls} for an example).}

\item{chr_participant}{A character string, the column name
for each participant's identifier.}

\item{lgc_progress}{Logical; if \code{TRUE} displays a
progress bar for the function run time.}

\item{...}{Additional parameters for the
\code{fun_to_fit_models} function.}
}
\value{
A list of lists, the model fit results for each
participant.
}
\description{
Function to fit models separately to each set of
data per participant.
}
\examples{
# Example data set
data(ddabcd_data_example_long)
dtf_delay <- ddabcd_data_example_long \%index\% list(1, 'eventname')

# Fit models using nls
lst_all_fits.nls <- ddabcd_fit_models_across_participants(
  dtf_delay, ddabcd_fit_models_using_nls, lgc_progress = FALSE
)

}
