% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_model_functions.R
\name{ddabcd_fit_models_using_tl_mle2}
\alias{ddabcd_fit_models_using_tl_mle2}
\title{Fit Delay Discounting Models via Maximum Tobit Likelihood}
\usage{
ddabcd_fit_models_using_tl_mle2(
  dtf_data,
  chr_outcome = "ddabcd_indifference_score_log_odds",
  chr_minus_log_delay = "ddabcd_delay_negative_log_months",
  chr_censored = "ddabcd_indifference_score_censored",
  num_limits = NULL,
  lst_priors = NULL,
  fun_priors = NULL,
  lst_start = NULL,
  lgc_warnings = FALSE,
  ...
)
}
\arguments{
\item{dtf_data}{A data frame with separate rows for each
delay duration and participant. Must have a column
with the log of the delay durations.}

\item{chr_outcome}{A character string, the column name
for the log-odds of the indifference scores to fit.}

\item{chr_minus_log_delay}{A character string, the
column name for the log of the delay durations.}

\item{chr_censored}{A character string, the
column name for the indicator of censored data.}

\item{num_limits}{A numeric vector with the lower
and upper limits below and above which data are
censored.}

\item{lst_priors}{An optional named list of vectors
specifying the hyperparameters for the prior on each
parameter.}

\item{lst_start}{A named list consisting of the elements
\code{'Null'}, \code{'M1987'}, and \code{'R2006'}.
Each list element should be a named vector with the
starting values for the parameter estimates (\code{'lnd'}
for all models, \code{'a0'} for \code{lst_start$Null},
\code{'nlnk'} for \code{lst_start$M1987}, and \code{'nlnk'}
and \code{'s'} for \code{lst_start$R2006}).}

\item{lgc_warnings}{Logical; if \code{TRUE} displays a warning
if any of the estimations failed.}

\item{...}{Additional parameters for the \link[bbmle:mle2]{bbmle::mle2}
function.}

\item{fun_prior}{An optional function that takes the
vector of parameters and a list of hyperparameters
to compute a prior on the parameters. Allows for
penalized maximum likelihood estimation.}
}
\value{
A list with the \link[bbmle:mle2]{bbmle::mle2} output for the null model,
Mazur's (1987) model, and Rachlin's (2006) model.
}
\description{
Fits several delay discounting models to a long-form
data set using the \link[bbmle:mle2]{bbmle::mle2} function from the
\code{'bbmle'} package and a tobit likelihood function
(CITATION) to deal with censored data.
}
\examples{
# Example data set
data(ddabcd_data_example_long)
# Take data from 1st participant
dtf_delay <- ddabcd_data_example_long[1:7, ]

# Fit simulated data for 1st participant in example data set
lst_fit.mle2 <- ddabcd_fit_models_using_tl_mle2(
  dtf_delay, control = list( maxit = 2e4 )
)

}
\references{
Mazur, J. E. (1987). An adjusting procedure for studying
delayed reinforcement. In M. L. Commons, J. E. Mazur,
J. A. Nevin, & H. Rachlin (Eds.), The effect of delay and
intervening events on reinforcement value (pp. 55-73).
Lawrence Erlbaum Associates, Inc.

Rachlin, H. (2006). Notes on discounting. Journal of the
Experimental Analysis of Behavior, 85 (3), 425-435.
https://doi.org/10.1901/jeab.2006.85-05
}
