% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_model_functions.R
\name{ddabcd_model_distributions}
\alias{ddabcd_model_distributions}
\title{Determine Model Distribution}
\usage{
ddabcd_model_distributions(chr_distribution)
}
\arguments{
\item{chr_distribution}{A character string,
the name of the distribution to use.}
}
\value{
A named logical vector with the
value for the corresponding distribution
equal to \code{TRUE}.
}
\description{
Function that returns a logical vector
indicating whether a normal, beta, or
logit-normal distribution should be used
for modeling purposes. Primarily for
internal use with estimation functions.
}
\examples{
ddabcd_model_distributions('Beta')

}
