% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_model_functions.R
\name{ddabcd_model_predictions}
\alias{ddabcd_model_predictions}
\title{Predictions and Simulations From Delay Discounting Models}
\usage{
ddabcd_model_predictions(
  num_param,
  num_delay,
  chr_distribution = "Normal",
  lgc_logit = FALSE,
  num_scaling = 1
)
}
\arguments{
\item{num_param}{A named numeric vector of parameter values.
Parameter names can be \code{b0} for the intercept-only
model, \code{a0} for the intercept-only model (logit-scale),
\code{k} for the discounting parameter, \code{lnk} for the
log discounting parameter, \code{nlnk} for the negative
log discounting parameter, \code{s} for the sensitivity
parameter (Rachlin, 2006), \code{d} for the dispersion
parameter, and \code{lnd} for the log of the dispersion
parameter. If \code{d} or \code{lnd} are provided,
function generates random deviates from the model, otherwise
it produces model predictions. The interpretation of
\code{d} and \code{lnd} depend on the specification of
\code{chr_error}.}

\item{num_delay}{A numeric vector of delay durations
(for \code{num_param['k']}), log delay durations
(for \code{num_param['lnk']}), or minus log durations
(for \code{num_param['nlnk']}).}

\item{lgc_logit}{Logical; if \code{TRUE}}

\item{num_scaling}{A numeric value, rescales results if
\code{lgc_logit} is \code{FALSE}.}

\item{chr_error}{The distribution for the dispersion around
model predictions. Options are \code{'Normal'}, \code{'Beta'},
or \code{'Logit-normal'}.}
}
\value{
A vector of predicted or simulated indifference scores.
}
\description{
Generates predicted or simulated indifference scores for
different delay discounting models. Supported models are
\enumerate{
\item a intercept-only model, 2) Mazur's (1987) model, and
\item Rachlin's (2006) model.
}
}
\details{
For an indifference score I scaled to be between (0 - 1),
Mazur's (1987) model is I = 1/(1 + k\emph{delay). Rachlin's (2006)
model is I = 1/(1 + k}(delay^s)). Mazur's model therefore is
a special case of Rachlin's when s = 1. The models can be
rewritten as a linear equation on the logit-scale:
logit(I) = -ln(k) - ln(delay)*s.
}
\examples{
# Delay duration in months
x <- c( 1, 3, 12 )

# Mazur's (1987) model
k <- 0.14
# Predictions given different parameterizations
ddabcd_model_predictions( c( k = k ), x )
ddabcd_model_predictions( c( lnk = log(k) ), log(x) )
ddabcd_model_predictions( c( nlnk = -log(k) ), -log(x) )

# Rachlin's (2006) model
s = .60
ddabcd_model_predictions( c( k = k, s = s ), x )
ddabcd_model_predictions( c( lnk = log(k), s = s ), log(x) )
ddabcd_model_predictions( c( nlnk = -log(k), s = s ), -log(x) )

# Random deviates
# Normal model
ddabcd_model_predictions(
  c( k = k, d = .3 ), x, chr_distribution = 'Normal'
)
# Beta model
ddabcd_model_predictions(
  c( k = k, d = 4 ), x, chr_distribution = 'Beta'
)
# Logit-normal model
ddabcd_model_predictions(
  c( k = k, d = 0.3 ), x, chr_distribution = 'Logit-normal'
)

}
\references{
Mazur, J. E. (1987). An adjusting procedure for studying
delayed reinforcement. In M. L. Commons, J. E. Mazur,
J. A. Nevin, & H. Rachlin (Eds.), The effect of delay and
intervening events on reinforcement value (pp. 55-73).
Lawrence Erlbaum Associates, Inc.

Rachlin, H. (2006). Notes on discounting. Journal of the
Experimental Analysis of Behavior, 85 (3), 425-435.
https://doi.org/10.1901/jeab.2006.85-05
}
