% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_plotting_functions.R
\name{ddabcd_plot_single_panel}
\alias{ddabcd_plot_single_panel}
\title{Plot Delay Discounting Data (Single Panel)}
\usage{
ddabcd_plot_single_panel(
  obj_data,
  obj_x = "ddabcd_delay_log_months",
  obj_y = "ddabcd_indifference_score_1",
  chr_xlab = "log(Delay - months)",
  chr_ylab = "Indifference score (Proportion)",
  num_xlim = NULL,
  num_ylim = c(0, 1),
  num_dim = c(width = 5, height = 5),
  int_pch = 19,
  num_lwd = 2,
  num_cex.pt = 1,
  chr_color = c("black", "black", "blue", "grey"),
  num_jitter = 0,
  chr_model = "M1987",
  num_margin = c(4, 4, 2, 0.5),
  lst_predicted = NULL,
  lgc_add = FALSE,
  lgc_new = FALSE,
  lst_prediction_settings = NULL,
  ...
)
}
\arguments{
\item{obj_data}{Either a data frame or a list of
model fits (see for example
\link{ddabcd_fit_models_using_nls}).}

\item{obj_x}{Either a character vector with the
column(s) to index for the x-axis values or
a numeric vector.}

\item{obj_y}{Either a character vector with the
column(s) to index for the y-axis values or
a numeric vector.}

\item{chr_xlab}{A character string, the x-axis label.}

\item{chr_ylab}{A character string, the y-axis label.}

\item{num_xlim}{An optional numeric vector with the
lower and upper limits for the x-axis.}

\item{num_ylim}{A numeric vector with the lower and
upper limits for the y-axis. Default is (0, 1).}

\item{num_dim}{A numeric vector specifying the
width and height of the plot panel when a
new plotting window is generated.}

\item{int_pch}{The type of point to draw.}

\item{num_lwd}{A numeric value, the width of the line.}

\item{num_cex.pt}{A numeric value, the size of the point.}

\item{chr_color}{A character vector giving the colors
to use for 1) the line for observed data,
2) the points for observed data, 3) the
line for predicted data, and 4) the
error bars for predicted data.}

\item{num_jitter}{An optional numeric value specifying the
limits for random uniform jitter noise to add to points.}

\item{chr_model}{A character string, the model to
extract results from if \code{obj_data} is a list
of model fits. Options include \code{'Null'},
\code{'M1987'}, or \code{'R2006'}.}

\item{num_margin}{A numeric vector of 4 values giving the
margin sizes for the bottom, left, top, and right edges
of the figure.}

\item{lst_predicted}{A list of either 1 or 3 elements
specifying either the columns or numeric values to
to use for \code{[[1]]} model predictions and
\code{[[2]]} and \code{[[3]]} the lower and upper
boundaries of the prediction interval.}

\item{lgc_add}{Logical; if \code{TRUE} adds new data
to an existing figure.}

\item{lgc_new}{Logical; if \code{TRUE} generates a
new plotting window.}

\item{lst_prediction_settings}{An optional named list
specifying how prediction intervals should be calculated
if a model fit object is provided. The list should contain
\code{lst_prediction_settings$width} for the width of the
prediction interval, \code{lst_prediction_settings$bootstrap}
set to \code{TRUE} for using a parametric bootstrap and
\code{FALSE} otherwise, and \code{lst_prediction_settings$samples}
with the number of samples to draw for the parametric bootstrap.}

\item{...}{Additional parameters to the \link[base:plot]{base::plot}
function.}
}
\value{
A plot.
}
\description{
Function to plot delay discounting data (typically
the log of the delay durations in months and
indifference scores as proportions) as a single
panel figure.
}
\examples{
\dontrun{
# Load in example data sets
data( ddabcd_data_example_wide )
data( ddabcd_data_example_long )
dtf_wide <- ddabcd_data_example_wide \%index\% list( 1, 'eventname' )
dtf_long <- ddabcd_data_example_long \%index\% list( 1, 'eventname' )

# By default data assumed to be long-form
ddabcd_plot_single_panel( dtf_long \%index\% 1 ) # 1st participant
# Wide-form data can be plotted by providing vector of column names
ddabcd_plot_single_panel(
  dtf_wide \%index\% 1, # 1st participant
  obj_x = ddabcd_data_delay_durations(),
  colnames( dtf_wide \%col\% 'indif' ),
  num_ylim = c( 0, 100 )
)

# Example using model fits for first participant
lst_fits.nls <- ddabcd_fit_models_using_nls( dtf_long \%index\% 1)
ddabcd_plot_single_panel( lst_fits.nls, chr_model = 'M1987' )
# Add predictions for Rachlin's (2006) model
ddabcd_plot_single_panel(
  lst_fits.nls, chr_model = 'R2006',
  chr_color = c('black', 'black', 'red' ), lgc_add = TRUE
)
}

}
