% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_model_functions.R
\name{ddabcd_post_fit_attr}
\alias{ddabcd_post_fit_attr}
\title{Extract 'ddabcd_attr' Component}
\usage{
ddabcd_post_fit_attr(obj_fit, chr_attr = "", obj_default = NA)
}
\arguments{
\item{obj_fit}{An R object for model fit results.}

\item{chr_attr}{A character string, the
specific attribute to access. Options include
\code{'predicted'}, \code{'residuals'},
\code{'rmse'}, \code{'loglikelihood'},
\code{'AIC'}, \code{'BIC'}, \code{'nobs'},
or \code{'dof'}.}

\item{obj_default}{Value to return if attributes
aren't found.}
}
\value{
A list of elements or a specific element
from the \code{'ddabcd_attr'} attribute.
}
\description{
Utility function to extract elements from the
\code{'ddabcd_attr'} component of the
attributes for model fits from
\link{ddabcd_fit_models_using_nls}
and \link{ddabcd_fit_models_using_mle2}.
}
\examples{
# Example data
data(ddabcd_data_example_long)
# Use data for first time point and first 3 participants
dtf_long <- ddabcd_data_example_long \%index\% 1:3
dtf_long <- dtf_long \%index\% list( 1, 'eventname' )

# Fit data using 'nls'
lst_all_fits.nls <- ddabcd_fit_models_across_participants(
  dtf_long, ddabcd_fit_models_using_nls, lgc_progress = FALSE
)

ddabcd_post_fit_attr( lst_all_fits.nls[[1]]$M1987 )['rmse']

}
