% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_model_functions.R
\name{ddabcd_post_fit_bootstrapped_prediction_intervals_mle2}
\alias{ddabcd_post_fit_bootstrapped_prediction_intervals_mle2}
\title{Bootstrapped Prediction Intervals for Models fit by mle2}
\usage{
ddabcd_post_fit_bootstrapped_prediction_intervals_mle2(
  obj_fit,
  int_samples = 1000,
  lgc_suppress_error = FALSE,
  chr_labels = c("a0", "lnk", "s", "lnd"),
  num_width = 0.95,
  num_limits = NULL,
  ...
)
}
\arguments{
\item{obj_fit}{A model fit object that has methods
for \code{coef} and \code{vcov}.}

\item{int_samples}{The numer of samples to draw
from the multivariate normal distribution.}

\item{lgc_suppress_error}{Logical; if \code{TRUE}
suppress error messages and return a matrix
of \code{NA} values instead.}

\item{chr_labels}{The column names for the matrix
of \code{NA} values if samples cannot be drawn.}

\item{num_width}{A numeric value between 0 and 1, the
width of the prediction interval.}

\item{num_limits}{An optional vector giving the lower
and upper limits below and above which data should
be censored.}
}
\value{
A 2 x N matrix whose rows give the lower and
upper bootstrapped prediction intervals for the N observations.
}
\description{
Compute bootstrapped prediction intervals via simulating
\enumerate{
\item possible parameter values from a multivariate normal
distribution, and 2) possible observed data values for
delay discounting model at each set of parameter values.
}
}
\examples{
# Example data
data( ddabcd_data_example_long )
dtf_long <- ddabcd_data_example_long \%index\% 1
dtf_long <- dtf_long \%index\% list( 1, 'eventname' )

lst_fits.mle2 <- ddabcd_fit_models_using_mle2(
  dtf_long, chr_distribution = 'Logit-normal'
)

ddabcd_post_fit_bootstrapped_prediction_intervals_mle2(
  lst_fits.mle2$R2006
)

}
