% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_model_functions.R
\name{ddabcd_post_fit_compute_rmse}
\alias{ddabcd_post_fit_compute_rmse}
\title{Compute Root-Mean-Square-Error (RMSE)}
\usage{
ddabcd_post_fit_compute_rmse(num_predicted, num_observed)
}
\arguments{
\item{num_predicted}{A numeric vector matching in
length to \code{num_observed}.}

\item{num_observed}{A numeric vector matching in
length to \code{num_predicted}.}
}
\value{
A numeric value, the RMSE.
}
\description{
Given a set of predicted and observed values,
computes the root-mean-square-error (also
known as the root-mean-square-deviation). Note
predicted values must be on the same scale as
the observed values.
}
\details{
Given \code{n} predicted values \code{p} and observed
values \code{o}, the RMSE is computed as
\code{sqrt( sum( (p - o)^2 )/n )}.
}
\examples{
# Example using 'lm'
set.seed( 20230517 ) # For reproducibility
dtf_data <- data.frame( x = rnorm(30) )
dtf_data$y <- dtf_data$x * .5 + rnorm(30)
lst_lm <- lm( y ~ x, data = dtf_data)
# Compute RMSE
ddabcd_post_fit_compute_rmse(
  predict(lst_lm), dtf_data$y
)
# Equivalent to...
sqrt( mean( residuals(lst_lm)^2 ) )

}
