% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_model_functions.R
\name{ddabcd_post_fit_convergence}
\alias{ddabcd_post_fit_convergence}
\title{Check if Models Converged}
\usage{
ddabcd_post_fit_convergence(lst_all_fits, chr_model = "M1987")
}
\arguments{
\item{lst_all_fits}{A list of lists with model fits.}

\item{chr_model}{A character string, the specific
model from which elements should be extracted.
Options include \code{'Null'}, \code{'M1987'},
or \code{'R2006'}.}
}
\value{
A logical vector equal to \code{TRUE}
if models converged, and \code{FALSE} otherwise.
}
\description{
Function to iterate over model fits across
participants and check for model convergence.
}
