% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_model_functions.R
\name{ddabcd_post_fit_extract}
\alias{ddabcd_post_fit_extract}
\title{Extract Elements From List of Model Fits}
\usage{
ddabcd_post_fit_extract(
  lst_all_fits,
  fun_extract = NULL,
  chr_model = "M1987",
  ...
)
}
\arguments{
\item{lst_all_fits}{A list of lists with model fits.}

\item{fun_extract}{A function that extracts a
specified element from the model fit (defaults to
the \code{coef} function).}

\item{chr_model}{A character string, the specific
model from which elements should be extracted.
Options include \code{'Null'}, \code{'M1987'},
or \code{'R2006'}.}

\item{...}{Additional arguments to the
\code{fun_extract} function.}
}
\value{
Output from the \code{sapply} function
after applying \code{fun_extract} to each
element of \code{lst_all_fits}.
}
\description{
Function to extract a specified element
(e.g., a parameter estimate) from a list of
lists of model fits (see
\link{ddabcd_fit_models_across_participants}
for an example).
}
