% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_model_functions.R
\name{ddabcd_post_fit_ifelse_null}
\alias{ddabcd_post_fit_ifelse_null}
\title{Call Function for Non-Null Model Fit}
\usage{
ddabcd_post_fit_ifelse_null(lst_fit, fun_call, obj_default = NA, ...)
}
\arguments{
\item{lst_fit}{A list with model fit results.}

\item{fun_call}{A function to apply to
\code{lst_fit}.}

\item{obj_default}{The output if \code{lst_fit}
is \code{NULL}.}

\item{...}{Additional parameters to \code{fun_call}.}
}
\value{
The results of \code{fun_call}, or if
\code{lst_fit} is \code{NULL}, \code{obj_default}
instead.
}
\description{
Utility function that applies a specified
function to a list with model fit only
if the list is non-null (i.e., successful
estimation).
}
\examples{
# Example data
data(ddabcd_data_example_long)
# Fit first time point and first and last participant
dtf_long <- ddabcd_data_example_long \%index\% c( 1, 48 )
dtf_long <- dtf_long \%index\% list( 1, 'eventname' )

# Fit data using 'nls'
lst_all_fits.nls <- ddabcd_fit_models_across_participants(
  dtf_long, ddabcd_fit_models_using_nls, lgc_progress = FALSE
)

# Estimation fails for Rachlin's (2006) model for last subject
# due to no variation in data to identify 2nd parameter
sapply( 1:2, function(i) {
  ddabcd_post_fit_ifelse_null(
    lst_all_fits.nls[[i]]$R2006, return(TRUE), FALSE )
} )

}
