% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_model_functions.R
\name{ddabcd_post_fit_report}
\alias{ddabcd_post_fit_report}
\title{Brief Report on Delay Discounting Model Fit}
\usage{
ddabcd_post_fit_report(lst_all_fits)
}
\arguments{
\item{lst_all_fits}{A list of lists with model fits
(see \link{dabcd_fit_models_across_participants}).}
}
\value{
A message to the console.
}
\description{
Function that provides a summary of (a) the number of
participants for which models where successfully fit,
and (b) means and uncertainty intervals for the
root-mean-square-error (RMSE).
}
\examples{
# Example data
data(ddabcd_data_example_long)
# Fit first time point and first 3 participants
dtf_long <- ddabcd_data_example_long \%index\% 1:4
dtf_long <- dtf_long \%index\% list( 1, 'eventname' )
lst_all_fits.nls <- ddabcd_fit_models_across_participants(
  dtf_long, ddabcd_fit_models_using_nls, lgc_progress = FALSE
)

ddabcd_post_fit_report( lst_all_fits.nls )

}
