% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_model_functions.R
\name{ddabcd_post_fit_sample_parameters}
\alias{ddabcd_post_fit_sample_parameters}
\title{Draw Samples For Parameter Estimates}
\usage{
ddabcd_post_fit_sample_parameters(
  obj_fit,
  int_samples = 1000,
  lgc_suppress_error = FALSE,
  chr_labels = c("a0", "lnk", "s", "lnd")
)
}
\arguments{
\item{obj_fit}{A model fit object that has methods
for \code{coef} and \code{vcov}.}

\item{int_samples}{The numer of samples to draw
from the multivariate normal distribution.}

\item{lgc_suppress_error}{Logical; if \code{TRUE}
suppress error messages and return a matrix
of \code{NA} values instead.}

\item{chr_labels}{The column names for the matrix
of \code{NA} values if samples cannot be drawn.}
}
\value{
A matrix.
}
\description{
Function that conducts a parametric bootstrap
assuming parameter estimates follow a multivariate
normal distribution.
}
\examples{
# Example data
data( ddabcd_data_example_long )
dtf_long <- ddabcd_data_example_long \%index\% 1
dtf_long <- dtf_long \%index\% list( 1, 'eventname' )

lst_fits.mle2 <- ddabcd_fit_models_using_mle2(
  dtf_long, chr_distribution = 'Normal'
)

ddabcd_post_fit_sample_parameters(
  lst_fits.mle2$M1987, int_samples = 3
)

}
