% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_model_functions.R
\name{dlogitnorm}
\alias{dlogitnorm}
\alias{plogitnorm}
\alias{qlogitnorm}
\alias{rlogitnorm}
\title{The Logit-Normal Distribution}
\usage{
dlogitnorm(x, mean, sd, log = FALSE)

plogitnorm(q, mean, sd, lower.tail = TRUE, log.p = FALSE)

qlogitnorm(p, mean, sd, lower.tail = TRUE, log.p = FALSE)

rlogitnorm(n, mean, sd)
}
\arguments{
\item{x, q}{Vector of quantiles
(\code{0 >=} \code{x},\code{q} \code{<= 1}).}

\item{mean}{Vector of means on the logit-scale.}

\item{log, log.p}{Logical; if \code{TRUE}, probabilities p
are given as log(p).}

\item{lower.tail}{Logical; if \code{TRUE} (default), probabilities
are P(X less than or equal to x) otherwise, P(X greater than x).}

\item{p}{Vector of probabilities}

\item{n}{Number of observations. If \code{length(n) > 1}, the
length is taken to be the number required.}

\item{prec}{Vector of standard deviations on the logit-scale,
where \code{sd > 0}.}
}
\value{
\code{dlogitnorm} gives the density, \code{plogitnorm} gives the
distribution function, \code{qlogitnorm} gives the quantile function,
and \code{rlogitnorm} generates random deviates. The length of the
result is determined by \code{n} for \code{rbetamp}, and is the
maximum of the lengths of the numerical arguments for the other
functions.
}
\description{
Density, distribution function, quantile function, and random
generation for the logit-normal distribution with mean on the
logit-scale equal to \code{mean} and standard deviation on on
the logit-scale equal to \code{sd}.
}
\details{
If \code{mean} or \code{sd} are not specified they assume the
default values of 0 and 1.

See \link[stats:Normal]{stats::dnorm} for more details.
}
\examples{
x <- c(.25, .5, .75)
dlogitnorm( x, 0, 1)
plogitnorm( x, 0, 1)
qlogitnorm( plogitnorm( x, 0, 1), 0, 1)
rlogitnorm( 3, 0, 1)

}
