% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_model_functions.R
\name{\%col\%}
\alias{\%col\%}
\title{Extract Columns Matching Partial Strings}
\usage{
dtf_data \%col\% chr_column_parts
}
\arguments{
\item{dtf_data}{A data frame.}

\item{chr_column_parts}{A character vector, the
partial strings to match column names against.
Columns that match any of the strings are returned.
However, if a string starts with \code{'+'}, all
columns must include that string. If a string
starts with \code{'-'}, no columns can include
that string. Note at least one element of
\code{chr_column_parts} cannot start with
either \code{'+'} or \code{'-'}.}
}
\value{
A data frame, the subset of columns meeting
the criteria specified by \code{chr_column_parts}.
}
\description{
Convenience operator that takes a vector
of partial strings and determines the columns
that match the strings. Also allows specifying
elements that must be included or excluded
when matching.
}
\examples{
dtf_data <- data.frame( A1 = 1:3, A2 = 4:6, B1 = 7:9, B2 = 10:12 )

# Return all columns with an 'A'
dtf_data \%col\% 'A'
# Return all columns with an 'A1' or 'B1'
dtf_data \%col\% c( 'A1', 'B1' )
# Return columns with a '1' and either 'A' or 'B'
dtf_data \%col\% c( '+1', 'A', 'B' )
dtf_data \%col\% c( '-1', 'A', 'B' )

}
