% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddabcd_model_functions.R
\name{\%index\%}
\alias{\%index\%}
\title{Index Rows of a Data Frame}
\usage{
dtf_data \%index\% obj_index
}
\arguments{
\item{dtf_data}{A data frame.}

\item{obj_index}{Either an integer vector, character
vector, or a list with the integer/character vector
and a character string specifying the column to
match over. If \code{obj_index} is an integer,
the operator uses the first column of \code{dtf_data}
to match over.}
}
\value{
A data frame, the subset of \code{dtf_data}
that had successful matches.
}
\description{
Convenience operator that matches a user-supplied
index (either an integer or character string)
against the elements in a data frame column and
returns the corresponding set of rows. Useful,
for example, to isolate rows for a specific
participant in a long-form data set.
}
\examples{
# Create example data set
dtf_data <- data.frame(
  V1 = rep( LETTERS[1:3], each = 2 ), V2 = c( 1:3, 1:3 )
)

# Return rows for first unique element of 'V1'
dtf_data \%index\% 1
# Return rows for first unique element of 'V2'
dtf_data \%index\% list( 1, 'V2' )
# Return rows for 'V1' == 'B'
dtf_data \%index\% 'B'
dtf_data \%index\% list( '3', 'V2' )

}
